# (#5429) - WIP - leveldb: simplify stored metadata json