@nolanlawson 2016-07-03
Marking this as WIP because I want to spend more time thinking about the impact of this in #5433. Changing the way we store data on disk is something irreversible.